xhc-hb04.tcl: use pre-existing signames
for required halui.* and motion.* OUT pins

If a [HAL]HALFILE executed prior to this file (LIB:xhc-hb04.tcl)
specified a signal name for a needed halui or motion output pin,
then use that name. (Prior behavior gave warning message and
continued).

Note: no attempt is made to accomodate existing signal names for halui
_input_ pins since a claimed input pin presumably means an output for
the signal is already made (or intended).  As before, a message is
is given and and execution continues.

Signed-off-by: Dewey Garrett <dgarrett@panix.com>