t0 tests: let introspection complete before continuing the test
The t0 tests have failed intermittently since they were written in 2012.
The failures has always involved missing lines of output, corresponding
to the MDI of user-defined m-codes that did not complete.  The fix is
to wait for introspection to complete before moving on to the next part
of the test.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>