axis.py: joint jog for special identity trivkins
as selected by jogging radiobuttons

The axis gui hides distinctions of joints and axes for
identity kins so when joint jogging (typ before homing),
it must determine a joint number based on the displayed
axis coordinate letter for the jogging radiobutton
selected by the user.

trivkins allows:
  a) omission of axis letters from the default sequence
  b) duplicate axis letters (gantry usage typ)

Before this commit, joint jogging failed for some identity
kinematics configs which explictly specified coordinates=
with omitted axis letters because the joint number used
was based on default coordinates for trivkins (xyzabcuvw).
Example (fails joint jogging for w):
    [KINS]KINEMATICS  = trivkins coordinates=xyzw
    [TRAJ]COORDINATES = XYZW

This commit selects the joint number for joint jogging for
trivkins with identity kinematics based on the letters
specified by [TRAJ]COORDINATES (local var is trajcoordinates).

Note: as documented (docs/src/getting-started/updating-linuxcnc.txt),
it is important that the coordinates specified with the trivkins
coordinates= parameter agree with [TRAJ]COORDINATES.

Note: for duplicate_coord_letters, the joint for the first letter
specified will be used.

Signed-off-by: Dewey Garrett <dgarrett@panix.com>