axis: fix preview when machine has locking rotaries
.. to better preview ungrounded gcode, axis issued a
"g53 g0 <coordinates>", which didn't work when there was a locking
rotary because it is not permitted to specify other axes to move
at the same time.

Instead, issue each move as a separate line, to satisfy the constraints
of locking rotaries.

This will work fine right up until the moment that the interpreter or
preview starts validating that the part program doesn't go outside
some complicated work volume ;-)

Signed-off-by: Jeff Epler <jepler@unpythonic.net>