Fix armhf builds so that required elements exist for comp OOT builds.
Out Of Tree builds have been broken for a long time on armhf packages.
Issue #1060 and others refer

The bottom of the problem is that the package machinekit-dev is built
with just posix flavour enabled, so as to avoid time constraints on
builds which otherwise would cause failure.

An unrealised consequence of this is that the file Makefile.inc,
which is pulled in by comp and instcomp, has flavours just set to
posix and the produced component targets posix, even if produced
on a rt-preempt or xenomai kernel.

machinekit-dev is largely an anachronism from linuxcnc days, when
in an effort to make the install as light as possible, all dev
stuff was in a seperate package.

This commit moves the -dev package files into the machinekit-<flavour>
package.  This ensures that the Makefile.inc contains settings which
match the targeted flavour.

Currently Debian Stretch will only build Machinekit, with local library
versions of czmq and zeromq, due to backward incompatibility of the new
versions.
This has a knock on in package building on Stretch, which fail dh_shlibdeps
tests unless a switch to ignore missing info is added.
Should be able to be removed once ported to new libs or our own ones produced.

Signed-off-by: Mick <arceye@mgware.co.uk>