packaging: support the 'nodoc' build profile
when using this build profile, the asciidoc documentation is not
built, and the build dependencies are reduced substantially.

Typical resource usage when using 'nodoc' profile:
    Need to get 181 MB of archives.
    After this operation, 555 MB of additional disk space will be used.
Compared to the baseline:
    Need to get 504 MB of archives.
    After this operation, 1408 MB of additional disk space will be used.

This mode can be enabled e.g., by setting the environment variable
DEB_BUILD_PROFILES=nodoc

More information on build profiles:
https://wiki.debian.org/BuildProfileSpec

In principle, our buildbot could be set to use this build profile
when building all but one architecture; however, there may be more
details needed before that can be done, such as the fact that some
manpages are now written in asciidoc, and these manpages are
included in the base package, not the disabled -doc packages.

Signed-off-by: Jeff Epler <jepler@unpythonic.net>