Motion: remove unused reading of [EMCMOT]COMM_WAIT
The userspace part of Motion (which is responsible for sending messages
to the main realtime part of Motion) reads an ini variable named
[EMCMOT]COMM_WAIT, and manages an internal variable holding the value,
but it is never used.  Remove it.

There is a short sleep in the "wait-for-ack-from-realtime-Motion" loop
of usrmotWriteEmcmotCommand(), but it's hardcoded to 25 microseconds,
it ignores this ini variable.

Signed-off-by: Sebastian Kuzminsky <seb@highlab.com>