Apply show_tool rotations in original GEOMETRY order
Impact is only for configs with more than one rotational axis.
Expectation is that rotations will be applied in the order of the GEOMETRY
attribute in the INI file. However when GEOMETRY is read in it is reversed.
This causes the tool rotations to be applied as -BA rather than A-B

Signed-off-by: Tom Schneider <cts@cs2corp.com>