axis.py: Dont try to convert unicode to unicode
I noticed when running in the ru_RU.UTF-8 locale that the message
  File "/home/jepler/src/linuxcnc/bin/axis", line 1419, in jogspeed_listbox_change
    if i == unicode(value, 'utf-8'):
TypeError: decoding Unicode is not supported
was displayed on the screen.  Almost certainly, the incremental
jog combobox would not behave properly when this error occurs.

Instead of making an assumption that 'i' is Unicode and 'value'
is 'bytes' (str), conditional reencode them based on their
runtime type.

Testing performed:
 * Test with LANG=C.UTF-8 and LANG=ru_RU.UTF_8
 * start axis, F1 F2
 * select continuous or incremental jog via keyboard
 * select continuous or incremental jog via mouse
 * with each selection, jogged X axis with arrow keys
No regressions detected.