hm2_7i43 hm2_7i90: Remove a kludge for crazy configs
The fix in 2.7 for the situation when the EPP cards were passed a blank
config string required a rather untidy workaround in case any existing
config was relying on the existing behaviour. The chances are almost zero
that anyone was relying on it so for master it seems appropriate to remove
the special handling.

Signed-off-by: andy pugh <andy@bodgesoc.org>