Fix spindle reversal calculations to avoid jitter and accelerations spikes.
Previously, the reversal was noisy because the TP would detect that the
spindle had reversed, and immediately set up a new motion segment for
the reverse motion. Unfortunately, if the current velocity is not zero,
the actual acceleration will be higher by v_final / dt.

Now, the reversal is smarter in several ways:

1) Store the direction and "origin" position of the spindle together in
one location in the TP. This way, the sign of the displacement depends
only on the commanded direction when the origin was chosen. It does not
depend on the current command / actual velocity.

2) Don't store a separate offset in the rigid tap struct for the
spindle. Whenever rigid tapping reverses spindle direction, update the
spindle origin in the TP.

3) Wait for the spindle to reverse direction by position (rather than
velocity). Also, wait for the current velocity to actually reach zero
before starting a reversal / final retract motion. This avoids the
acceleration spike.