Refactor limit3 to resolve overshooting and drifting
Also, add velocity constraint checks in tests.

There's some source of error that makes it difficult to stop exactly
before the limit, and there can be a small overshoot, less than 1% in
the test example.  The tests' limit checks are fudged to accomodate
this.  After beating my head at this for quite some time, I've decided
to chalk it up to JMK's warning about "the discrete time nature of the
code vs continuous time of normal physics, where V = integral(A) and P
= integral(V)" [1].

[1]: https://github.com/LinuxCNC/linuxcnc/issues/240#issuecomment-337326649