Use Python's logging module instead of prints
print is not good for debugging/logging because unless you spend the time
to put a lot of detail into each print line, it does not give you enough
information.  Also since there is no way to `silence` debug prints it is a
temptation to limit their use, or even remove/comment them out, which can make
finding a problem later-on much more difficult.

This commit adds facilities for using Python's logging module to log to both
the terminal and a log file.  A subclass of `logging.Formatter` is used for the
terminal handler to produce colored log levels, and optionally allow tagging
text in the log message to colorize it as well.