Smart Serial: Start to enable parameter setting via HAL
The only way to change settings on the smart-serial remotes is currently
with setsserial. This makes some sense with non-volatile parameters, but
is not useful for (for example) changing the position scale on a servo
drive.
This commit makes all smart-serial remote parameters visible
to HAL and settable using the normal "setp" command (the "setsserial
utility is no longer required for this)
Parameter changes only take place when the associated smart-serial port
tansitions from the "stopped" to "run" state.

Signed-off-by: andy pugh <andy@bodgesoc.org>