qtvcp -More Focus overlay and dialog work
We are begining to win here I think
After MUCH work it was discovered that the problem
of the overlay not moving past the visible screen was because
the window flag was set as a QT.Tool widget which must not go
off screen. After a happy accident it was discovered that when
the TOOL flag was removed then it was not nessasary to move the
overlay over the window. Now it just seems to work.
There are some keypress events to trap yet but moving forward.