axis.py,.tcl fixes for 285,361
For the master branch, the axis gui:
  a) attempts to allow alternate jogging methods (wheel jogging,halui)
     when the mdi tab is selected.
  b) avoid bad behavior observed at commit: 4f90341177c4204dd315

This commit attempts to address cooperating guis that issue
mdi commands as noted in issue 361 and the transient pulsing of
halui.mdi.is_manual when selecting the mdi tab as noted in issue
285.

If the axis mdi tab is active *and* an mdi command has been
issued, cooperating guis that issue jog commands must set to
manual mode (pulse halui.mode.manual for instance).

Note:

The behavior for axis gui tabs was changed in 2.8~pre (master)
because it was inconvenient for users to try to wheel jog when
the axis gui tab was at mdi.  With this commit, it will be
possible but user ancillary guis will need to address setting
mode to manual if an mdi has been issued.