Add note to emcled.cc re Stretch onwards
Appears not worth the effort of converting this to use ip a and ip route
to replace ifconfig and netstat respectively.
The output is different and would require a re-write of something
it is unlikely anyone uses.

The package net-tools is still available, so test for netstat and ifconfig
and warn users that they need to install that package if not found.