Prevent fail when doing item assignment later.
Trying to resolve #328, succeeded.
When one of rotary axis is wrapped, the code between line 1292 and 1302 will do item assignment in positions.
But if self.get_show_relative() returns false, positions is a tuple, item assignment will fail. So I added 2 lines of code to ensure positions is always a list.