limit3: add .in-limit pin
This pin was present in the master version of limit3 before the merge of
the fix for issue #240 (but it was not in the version of limit3 in 2.7,
so nothing's needed there).

The .in-limit pin is True when the limit3 output is prevented from
tracking the input by limits (position, velocity, or acceleration).  The
.in-limit pin is False when the output is tracking the input perfectly.