qtvcp -macro: add software entry keyboard
one can use a touch screen to enter data.
This should probably be an option as it would be annoying
if using a keyboard.
This software keyboard should probably have a conversion button from
mm to inch or vice versa based on current active units
possibly nuttons for current axis position too...