limit3:  improve math for stopping at limits
Max/min limits were previously overshot by up to 1% in tests because
the math was slightly off, and also because the logic didn't
understand that a little blip in acceleration could cause limit
overshoot.

This commit breaks the soonest-possible stop position calculation into
two parts, one for most positive acceleration during this period, then
stop; another for most negative acceleration during this period, then
stop.  This forms a moving window that makes it easier to detect when
we'll be stopping too early or too late.