limit3:  avoid max/min overshoot from locking onto limits too early
When approaching max/min limits, even if the limit *position* can be
reached in the next period, it may not be possible to actually *stop*
on it without violating acceleration limits.

This patch detects that condition and instead accelerates away from
the limit.  This allows the tests' 0.1% "fudge factor" to be
eliminated.

However, this can also create a situation where the output endlessly
repeats a cycle of accelerating toward the limit, seeing it is going
too fast, and accelerating away.