Change all instances of `/usr/bin/env python2` to `/usr/bin/python2`
Whilst the two shebangs atop python scripts are interchangable in Debian currently,
there is no knowing what using `env` to determine the interpreter would result in
on other distros that for instance are not Debian based.

In the very worst scenario, this means that creating a single symlink in
`/usr/bin/python2` will enable ALL python scripts to work on any distro.

(Thus far not aware of a distro which does not already have that symlink,
as part of the python2.7 installation.)

Signed-off-by: Mick <arceye@mgware.co.uk>