Fix runtests for Stretch and Buster
Quite a few runtests rely upon linuxcncrsh used in conjunction with
`tcping` or `nc`

The netcat package containing these binaries has been deprecated,
starting at Stretch, along with a lot of the old net-utils.
(Even ifconfig is not available unless specifically installed,
from the package net-tools)

Add a `skip` script to those tests which require deprecated binaries
to test for presence and skip if not found.

Not doing so will cause the failure of that and every subsequent
test because the linuxcnc/halcmd processes are left running.

Signed-off-by: Mick <arceye@mgware.co.uk>