qtvcp -fix a seg fault with simple screens
screens that didn't use the graphics program would
segfault because only the graphics program had
the command gobject.init()
I'm not sure why qt5 requires this but it seems to work.
Now hal_glib will not do this because it causes a seg fault.
but qt_glib overides this function.