limit3:  Fix acceleration
The input signal or a max/min limit would be locked onto too early,
and while position and velocity limits were satisfied, acceleration
limits would be violated.  This bug was reported in #378.

Fixing this problem revealed that always using maximum acceleration in
either positive or negative directions could oscillate back and forth
before achieving lock, and in some cases would never lock and would
oscillate forever.

This patch resolves the problem by smoothing out acceleration during
approach so that when the output position reaches the input (or
max/min limit), it can lock on without hunting around and without
violating acceleration limits.

Because the update function only runs periodically, the approach must
be planned so that the lock-on timing coincides with an update.  This
means that the approach may be at slightly less than maximum
acceleration.  The functionality tests check for less-than-maximum
acceleration, and this causes them to fail.  Also, because there is no
`modf(3)` function available, planning the approach to an integral
number of periods is done by casting a double to an int and back; this
will break in the (unlikely) case of a very long approach.

These problems are solved by smoothing the acceleration only in the
last n periods, by default two.  This variable is broken out into a
pin so that it may be set differently for other applications.  A
higher setting will smooth out acceleration blips commonly seen at
greater distance, an effect resulting from the error of periodic
updates and the lessening velocity difference allowing the approach to
be refined.  For applications where these blips are undesirable, the
`smooth-steps` pin should be set to a value larger than the number of
periods of any possible approach.

This patch also improves the math so that the max and min limits are
never overshot.  Previous commits updated the tests to reflect this.

This patch also improves the math over PR #380, where the velocity
calculation was done differently, causing the first period to
accelerate at 1/2 maximum.

Fixes #378