rtapi: fixup signed int min/max constants
This fixes a problem with RTAPI_INT32_MIN on Wheezy/RTAI platforms
(and possibly others) that led to an immediate hostmot2 stepgen runaway
at power-on.

A new build-time warning was introduced when RTAPI_INT32_MIN was added recently:

    warning: this decimal constant is unsigned only in ISO C90

This warning showed up in the 1401.rip-wheezy-rtai-i386 buildslave's
build 3119 (the new stdint branch) and 3120 (master with stdint merged):

http://buildbot.linuxcnc.org/buildbot/builders/1401.rip-wheezy-rtai-i386/builds/3119
http://buildbot.linuxcnc.org/buildbot/builders/1401.rip-wheezy-rtai-i386/builds/3120

It was not present in the previous build of master:

http://buildbot.linuxcnc.org/buildbot/builders/1401.rip-wheezy-rtai-i386/builds/3113

When building non-rt (userspace) code for RTAI we explicitly set the gcc
std to gnu99, but when building kernel code for RTAI we don't set -std via
the cflags, so on wheezy's gcc 4.7.2 we get the default, which is gnu90.

This stackoverflow discussion explains the problem:

    https://stackoverflow.com/questions/9941261/warning-this-decimal-constant-is-unsigned-only-in-iso-c90

This commit changes rtapi_stdint.h's definition of RTAPI_INT32_MIN to
match the systemwide /usr/include/stdint.h on Wheezy and Stretch (and
probably others).