packaging: improve short descriptions
The "short description" of a package is the first line of its Description
field in the Debian Control file.

This commit updates the short descriptions of our binary packages to
be different from each other, and hopefully more descriptive/helpful,
as preferred by lintian.

Also change the wording slightly to be more consise and accurate.