packaging: validate desktop files
Version 0.22 and newer of the desktop-file-validate program (in Jessie
and newer) can deal with multiple files specified on the command line.

Version 0.20 (in Wheezy) says it can, but can't.  So feed it the files
one by one.

Also fix numerous problems with the desktop files that crept in over
the years.