docs: fix hyphen/minus confusion in manpages
groff treats '-' (the character you get when you hit the "minus" key on
the keyboard) as "hyphen", not as "minus".  Thus it renders incorrectly
in some locales, and line-wraps strangely.

groff treats the two-character sequence "\-" as "minus", and the
four-character sequence "\(hy" as "hyphen".

Details here: https://lists.debian.org/debian-devel/2003/03/msg01481.html

This commit replaces every instance of "-" in our manpages where the
meaning is "minus" with "\-", so it works right.

This fixes many lintian warnings.