io: set the HAL pins/params even for the loaded tool
This fixes #330.

Before this commit, this sequence of G-code commands led to incorrect
behavior:

    N1 T1
    N2 M6
    N3 T2
    N4 T1
    N5 M6

The T2 on N3 would set io's .tool-prep-number and .tool-prep-pocket
pins, but the T1 on N4 would not, so the pins would keep their old
T2 values.  The M6 on N5 would incorrectly load T2 instead of T1.

The same problem would arise in a program without N4, if the program
was aborted after N3 and before N5.

This commit fixes the problem by setting io's hal pins even for the tool
in P0.