lcd.c - fix a read out of bounds bug
An off-by-one error meant that the first character from the next page was
always appended at the end of each page. And, in the case of the last page,
the contents of random memory would be displayed.
This patch also fixes some formatting issues with escaped characters,
makes the screen clear properly at initialisation and allows the decimal-
separator HAL parameter to actually work properly.

Signed-off-by: andy pugh <andy@bodgesoc.org>