halmeter -add scroll-to-keypressed letter search
if you keypress a letter while in the selection window,
halmeter will scroll to the first name that starts with that letter.
It is currently case sensitive.

It doesn't actually select the name as I couldn't figure out how to
get the cursor position to follow the selection. If you keypress picked
a name then used the cusor keys it would jump back to the previous selection.

If we update to current GTK code, could probably fix that - in the future