canon:  fix feed rate problem in arc corner case
When an arc is 360 degrees, the endpoint and the starting point are
the same.  Naively calculating the XYZ movement using only the
difference between the endpoints overlooks XYZ motion.  When such a
move is combined with a rotary axis movement, the resulting feed rate
will be incorrect.

This patch fixes the problem by passing a parameter to the function
calculating that distance that forces XYZ movement to be recognized.

This may not address the entire issue, since later code in that
function makes further use of the endpoint differences.