Test rtapi_task_pll_set_correction()
This tests rtapi_task_pll_set_correction() by setting a target,
running 2000 cycles, averaging error from the last 1000 cycles, and
testing that error is less than 1% of a period.  This is intended to
be liberal enough to pass even on posix threads.

The first target is 0, the second target is 10 periods in the future.