qtscreen -fix seg fault when loading file
I have no idea why this fixes the seg fault.
It seemed to be a race situation because some times it did not
seg fault. If it didn't the first time it didn't the rest of the session.

This seg fault seems to have showed up after the change to add qtvp.core
which consolidated other libraries.

here is hoping this is a permanent fix