Proposed fix for "FIXTHIS too verbose"
I don't like the idea of having two different styles
for the G-code and M-code tables, but there are just
too many G-codes available to not have it be verbose.
So I ask the user to pick the blank cells for G-codes.

I could "reverse" the M-codes table too, if consistency
between them is thought to be important.

The tables are broken up to avoid trying to print
one table 100 lines long.

Signed-off-by: Kim Kirwan <Kim@KimKirwan.com>