configs -add a control panel for testing TOUCHY
It's pretty useless as a sim or a test screen if one can't actually control it.
We add a test control panel with warning label that it's just for testing.
It would be really annoying anways as it sitd on top of the screen and needs to be
moved around to see everything.

This panel just has basic controls for jogging and cycle start.
I've only added it to this one sim for now.