tram.c: Fix pointer arithmetic thinko
Because the type of tram_read_buffer is uint32_t, it's not
appropriate to perform arithmetic on it with the old read size
which is in bytes.

This didn't usually affect hm2 EXCEPT FOR bspi, because bspi
is the only user of the code that reaches hm2_allocate_tram_regions
with the old_tran_read_size being nonzero.

(The analysis is just the same for the case of the write buffers)

For those following along at home, I added the bad code
in cf4f68997553422c40dea22ad916b82ff0254a7f and all 2.7 and master
branches are likely affected.

Closes: #451