Configs comments: Avoid single quotes/apostrophes
When using the syntax highlighter in gedit, single quotes and
apostrophes cause the highlighting to go bright magenta and
stay that way until a matching single quote or apostrophe is
found.

It would be nice to think that this could be fixed by adjusting
the syntax highlighter, but since that has not been done by the
experts, it's probably a tricky job. I think we *do* want the
magenta highlighting to appear in spots "like this" and 'like
this', and I'm not sure what it would take to make that work.

So I'll just go ahead and fix the offending config files.

The old joke says it best:
Patient: Doctor, it hurts when I do *this*.
Doctor: Don't *do* that.

Signed-off-by: Kim Kirwan <kim@kimkirwan.com>