scripts/realtime:  check for zombies
When running in a container, there is no proper `init` process to
adopt the orphaned `msgd:x` or `rtapi:x` daemon processes, so when
these daemons exit, they are left behind as defunct processes.

This isn't inherently a problem, except the realtime script uses
e.g. `pidof rtapi:0` to check for a running daemon; however, this can
give a false positive.

This patch introduces `FindRunningProcs $proc_name` and
`CheckRunningProcs $proc_name` to filter out zombie processes and
eliminate those false positives.

It may also be possible to fix this properly with some scheme to fork
off a background process that runs the daemons with the `-f` flag, but
in the background, and then use `wait` to properly reap the processes.
However, running in containers is not a primary use case, so this
simple workaround may be sufficient.