tp: Fix fallback to parablic blending when tangent blend fails.
Fixes #447, which was caused by the TP not correctly falling back to
parabolic blending in certain cases. There is now logic to choose either
arc blending, treating segments as tangent, or parabolic blending,
depending on which is estimated to be fastest. This way, we should avoid
corner cases, since the performance check is always done when a new
motion is queued.