WIP: Interp: put G38.n absolute probed position in #5081-#5089
NOTE: This is a work in progress.  Only emccanon is properly updated.
The other canon implementations need matching updates.  Also I'm not
sure if this belongs in 2.7 or master.

The existing parameters #5061-#5069 currently (before this commit)
hold the probed position in the work coordinate system.

This commits stores the same probed position but in the absolute machine
coordinate system (G53) in #5081-#5089 (these parameters were unused
before this commit).

The intent is to help set the work coordinate system using automatic
probing of the part, without having to care which work coordinate system
is currently active:

    ; set X=0 at left edge of part
    G38.2 X10
    G10 L2 P0 X[#5081 + #<probe_radius>]