pyvcp: let the user specify number format in Bar widget
Before this commit, the number shown in the Bar widget was always printed
with one number after the decimal place ("%3.1f").

This commit makes the number format configurable by the user, using
this syntax:

    <bar>
        <format>"5.3f"</format>
    </bar>

The value of the format element defaults to the old built-in value of
"3.1f" if omitted.

This new Bar element duplicates the existing syntax for other PyVCP
widgets that have configurable numeric format.