docs: fix command for launching pyvcp example
Jin^eLD reported on IRC that the command to launch the tiny.xml  PyVCP example did not work. If using halcmd instead of `halrun -I` it does work. `halrun` with the -I flag should be the same as `halcmd` (if I understand it correctly) but it does not seem to work. This may be a bug in `halrun` ...