Provide a clearer error message when exceeding hal pin name length
The maximum length of a pin name is calculated after an internal
construction of the string, which is not obvious to the user. The error
message will now print the constructed string, so that it becomes clear
why a "component_name.pin_name" combination, that is still shorter than the
maximum allowed size, fails.

Signed-off-by: Sergey 'Jin' Bostandzhyan <jin@mediatomb.cc>