Update debian/control.xxx.in files
Situation at 20092018
~~~~~~~~~~~~~~~~~~~~~
python-gst1.0 backported to Jessie + Stretch
This affects gmoccapy, which uses functions in the old lib, but checks were introduced in
https://github.com/machinekit/machinekit/commit/5e5bb64 to remove those if newer distro used.
These remain valid as python-gst-0.10 is not available for Stretch and above even though the
newer lib is available in Jessie.

python-gtksourceview2 is still not available for Buster, but Sid has it, so expect backport soon
This is used in gladevcp widgets.

python-pil and python-pil.imagegtk active for Sid replacing python-imaging and python-imaging-tk
however selecting python-imaging and python-imaging-tk in Stretch actually
results in python-pil and python-pil.imagegtk being installed, so will probably completely backport soon.

Keep individual control files, so they can easily be amended as things change

Signed-off-by: Mick <arceye@mgware.co.uk>