gladevcp - hal_mdi_history - make it react to key events
Now a key up or down will get the marked value to the entry,
it can than be modified directly without using the mouse to get
the focus back to the entry.

hal_mdi_history will now react equal to axis mdi history.

Additionally I added a feature, that the last three items will not
be added again if that entry's has not been changed.

Signed-off-by: Norbert Schechner <nieson@web.de>