pncconf -remove signal to PID.command-deriv
I was advised by Peter of Mesa that using PID's own cmd-devir
is better then using an external one.
Even worse is if a signal is connected but no pin is, then
FF1 doesn't work at all, because the PID uses black magic to know
if the pin is connected to a signal to choose behaivor.
So now we don't ever put a signal to it.