Makefile:  add `CPPFLAGS` to RT `gcc {%.c,%.cc}` command
While updating the Docker images for Jessie CI to replace the emdebian
cross-compile toolchain with Linaro's, the RT build mysteriously
broke, and I'm unclear why the behavior should be different.  (I also
wonder if this would obviate any of the monkey-patching in the
Dockerfile.)

This is related to revs. 0b06a453 and de1ad060 from
PR machinekit/machinekit#1077 and issue machinekit/machinekit#1072