add ~/.local/bin to PATH in linuxcnc start script
There is a regression in bash shipped with Debian Stretch which results in `~/.local/bin` not being on the path. This is the location to which pip installs python entry points, so is needed for linuxcnc to be able to find and launch some python based displays.

Thanks to JT for writing the needed lines of bash.