tp: Fix handling of approximate tangents and required de-rating of max velocity and acceleration.
Instead of directly altering the segment's maxaccel field (which is not
re-entrant), store an acceleration reduction parameter for
approximate tangents. This way, the actual max acceleration can be
computed on-demand (knowing the original max acceleration and de-rating
due to approximate tangency).

A nice side-effect of this is that the reductions are not applied twice,
even in the code paths do the tangency check twice. This squeezes
slightly more performance out of the blends.