tp: Set flags on next segment when setting terminal condition.
The "blend_prev" flag needs to be set on the next segment whenever the
current segment uses parabolic blending (i.e. TC_TERM_COND_PARABOLIC).

Eventually, we should redesign functions that use blend_prev to look
directly at the previous segment's end condition. Using this flag as we
do is a potential source of bugs because it has to be changed any time
the terminal condition does.