tp: Refactor circle maximum velocity / acceleration ratio calculation
Slightly redesign the acceleration ratio / maximum velocity calculation
for circular motions. For circles with small radii relative to their
requested velocity, limit the tangential acceleration to 50% of the
peak, and cap maximum velocity so that normal acceleration is 86% of
peak.

For large radius circles, allow the tangential acceleration to be
higher, since less of the acceleration budget will be used for normal
acceleration.