tp: Refactor and simplify parallel / anti-parallel checks
In several places in the TP, we need to check if unit vectors are
parallel and / or anti-parallel. Previously, this was implemented with
dot product / arccos, which could be expensive to evaluate. This is now
implemented as a tolerance on the delta between the sum or difference of
the vectors, which is fast, and numerically robust (no divisions /
trancendentals).