tp: Overhaul debug output with JSON5-formatted log statements.
Provides a simple API to generate JSON5-formatted logs from within TP
code. This approach is crude in that the JSON string is built up in
place with a sequence of statements, and must verify that the log output
is what they expect.

A typical TP log statement now looks like this (all on one line):

{ log_entry: "SomeEntryType", log_level:"MyLogLevel", field1: 123.456,
...}

This is an ugly solution that abuses the logging system to dump internal
state for debugging. Efficiently packing / exporting internal state
would be much more efficient with flatbuffer or protobuf (which both
have C implementations). Luckily, the debug code is ignored entirely in
release builds due to the prepreocessor flags.

There is also an example of a python script that shows how to use a
JSON5 library to extract TP state from the debug log and plot it.