rs274ngc_pre.cc:  Fix MDI calls after sub updated
As long as it stays in MDI mode, interp never clears out the
`offset_map` cache of `O` offset pointers.  Calling a sub in MDI
causes its `O` offsets to be cached.  Editing the sub to change
the `O` offsets within the file and calling the sub a second time
results in an error:

    unwind_call: call_level=1 status=INTERP_ERROR - error: 5 from execute emc/rs274ngc/rs274ngc_pre.cc:516
    unwind_call leaving sub ''
    unwind_call: setting sequence number=11 from frame 1
    unwind_call: exiting current sub ''
    unwind_call: call_level=0 status=INTERP_ERROR - error: 5 from execute emc/rs274ngc/rs274ngc_pre.cc:516
    unwind_call: call_level=0 status=INTERP_ERROR - error: 5 from execute emc/rs274ngc/rs274ngc_pre.cc:534
    <-- execute(): error returned, clearing remap and call stack
    emc/task/emctask.cc 389: interp_error: 5: duplicate O-word label - already defined in line 4: 'o100 if [ #1 GT 5]'
    5: duplicate O-word label - already defined in line 4: 'o100 if [ #1 GT 5]'

This patch clears out the `offset_map` after executing an MDI block,
forcing subs to be reread on the next call.