AXIS -add ability to set hal.ready() in .axisui
I personally don't think this is a great idea - so I won't document it.
It can hide hal connecting code that usually should be in a HAL file in the .axisui file.
This could easily frustrate the typical user.
But for really customized AXIS screens built by power users - this can be convenient.