tp: Use already-calculated max normal acceleration from canon for optimal circle max velocity in TP
This is a small API change to the TP to pass in some additional calculations
that canon has already done. Basically, the TP needs to know more about
what the acceleration limits are so it can scale down the maximum value
correctly. For high-slope helices, the effective radius is quite high,
so the normal acceleration has a much smaller effect. Previously, the TP
had no easy way of knowing how much of the acceleration limit was due to
the helical axis limit, and how much was from the axes in circular
motion.