tp: Add position mismatch check when starting a motion segment
This adds a check to prevent "drift" in TP's position output. When a new
segment is activated, check the TP's current position against the
initial position of the new motion segment. If they don't match, then
the TP has failed to execute the previous path, and arrived at the wrong
position.

This should only occur if there is a serious regression in the TP, but
it will be useful during testing to catch dumb mistakes. One simple way
to cause such a bug is to fail to abort creation of an arc blend after
the previous / next segments have been shortened (to make space). The
machine will very quickly and very drastically deviate from the expected
position as a program executes.