qtvcp -fix multiple dialogs showing up
Since each button instance had it's own dialog, a call to show the dialog
would show a dialog for the total of axis_tool_button instances.
Now we must add a dialog to the main ui and all the toolbuttons will use that one.
Modify the screens that use axistoolbuttons to have a dialog in the main screen.