qtvcp -screen options for loading dialogs and focus overlay
It's much easier to select dialogs here then place them on the screen.
Since they are usually hidden and the show posiiton is controlled by
the preference file - placement on the screen is unimportant.
i should add the rest of the dialogs..TODO

The focus overlay is really difficult to place and have it work properly.
The seems to work reall well - and is more convenient.
There are some options not exposed such as image paths...TODO