test: Functional tests for TP from NGC snippets known to reproduce past issues
This adds de-facto regression / functional tests for the TP. The
batch-test runs a small set of NGC programs that have previously
revealed bugs in TP / motion, and looks for any reported errors,
including:
* Violations of any axis acceleration, velocity, or position limits
* Position "drift" in the trajectory planner (i.e. mismatch between TP
and motion commands)

The batch test provides a standard test runner for the functional tests, as well as
error / result checking. This also replaces a lot of the hacked-together
test functionality in the "circular-arcs" test folder (which was not
designed for standalone / automated testing).

Note that the circular-arcs folder still has utilities for manual
testing, but a lot of the NGC files have been moved over to the batch
test directory.