canon: Use enums instead of typedefs and defines
C++ makes #defines for const signed int values all but obsolete, and
canon.hh is a pure C++ header. The same is true for the "typedef struct {...}
StructName;" pattern, which is an anti-pattern in C++.

Also, propagate this change through to various places (like python
modules) to avoid changing the external API. Canon enum values are
exported at module scope, for example.