unit_test: add hello world interp unit test and better meson build infrastructure
The meson build system is expanded a bit:
* meson.build files in subdirectories list files / include directories
(relative to that subdir) for later use. These build files should NOT
include any dependencies (so they can be included in any order)
* top-level meson.build specifies everything else (deps, libraries,
targets, etc.) after including the subdir files
* Some kludgy mock files are used to pave over stray missing symbols not
really needed for unit tests

The meson build could eventually be expanded to replace the existing autotools / make build if
desired (there are guides on how to do this), or scrapped and replaced
with CMake 3 if need be.