packaging: buster needs the py2 python-yapps package
Our packaging runs our halcompile tool, which is a python2 program that
imports yapps.

Our package build-depends on yapps2, which on older (pre-Buster) distros
pulls in the correct python2 yapps runtime package, but Buster switched
the yapps2 runtime to the python3 version.

This commit updates debian/configure to, on Buster, explicitly
build-depend on the py2 python-yapps package.

The halcompile runtime dependency (for the linuxcnc-dev package) uses
${python:Depends} to automatically depend on python-yapps, so no change
is needed there.