Adds PORT pin type to hal.
A HAL port pin is a hal pin that acts as a real time one way byte oriented data stream An output port on any component may be connected to an input port on any other component via a signal. Data written on the output pin becomes accessible to the input pin. A HAL port signal may link only a single writer and a single reader. A port buffers data with a user definable buffer size.

Includes:
	Documentation on how to use a port object.
	a ctypes wrapper for the hal api (pyhal.py)
	a test to exersice hal port code along with pyhal.py
	changes to allow halcmd to understand and manipulate ports.