qtvcp -update istat: addtool, hash curent
add a tool to the front of the array so it's easier to find.
add bool return values and change the same of the function
to compare the hash value between now and last.
It the function retrns false we need to update things.
Switch from calling linuxcnc directly to using ACTION library.