qtvcp -tooloffsetview: fix conversion of toolinfo
we use a trick to cover tool wear for lathes.
but we need to convert from linuxcnc and qtvcp info arrays.
we were not adding some of the information so somtimes
tooloffsetview would not register the change on screen.

Switch to a ascending tool number view.
move code for future upgrade to metric/imperial conversion.