qtvcp -revamp toolfile staleness check, add delete tool
previously it was easy for multiple widgets reading the HASH code to
fool each other that the toolfile hadn't changed.
Now the toolfile is checked with STATUS and then if it has changed
we emit a signal with the new data, again using STATUS

Also added a delete tool function that takes a list of tool numbers to
delete. It loads and saves the tool file, ignoring the tools in the list.

added a toolfile-stale signal to STATUS to facilitate this. This signal
is not available in GStat (gladevcp's version)

Updated tooloffsetview widget to use new signal for update